fix: add cast to double in get time balance method

+++ src/main/java/com/compartetutiempo/timebank/member/Member.java
@@ -51,7 +51,7 @@ public class Member extends Person {
 
     // TODO: Debe estar entre -5 y 10
     public Double getTimeBalance() {
-        return hours + minutes / 60.0;
+        return hours + (double) minutes / 60.0;
     }
 
 }
